+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
+## [unreleased]
+
+### Added
+
+- `ui/ImageItem` prop `slotBefore` to add a node before the image
+
 ## [5.0.0] - 2025-06-13
 
 No significant changes.